[Fix] Wrong type at PoiTextStyle.

+++ lib/models/label/poi_text_style.dart
@@ -2,16 +2,16 @@ part of '../../flutter_kakao_map.dart';
 
 class PoiTextStyle {
   final double aspectRatio;
-  final double characterSpace;
+  final int characterSpace;
   final int color;
   final String font;
   final double lineSpace;
-  final double size;
+  final int size;
   final int stroke;
   final int strokeColor;
 
   const PoiTextStyle({
-    this.aspectRatio = 0,
+    this.aspectRatio = 0.0,
     this.characterSpace = 0,
     this.color = -16777216,
     this.font = "",
